requirements.txt
absl-py==0.13.0
astroid==2.5.6
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
click-plugins==1.1.1
cligj==0.7.2
cycler==0.10.0
decorator==4.4.2
Deprecated==1.2.12
dimod==0.9.13
dwave-cloud-client==0.8.4
dwave-greedy==0.1.2
dwave-hybrid==0.6.1
dwave-inspector==0.2.5
dwave-neal==0.5.7
dwave-networkx==0.8.8
dwave-ocean-sdk==3.3.0
dwave-qbsolv==0.3.2
dwave-system==1.4.0
dwave-tabu==0.3.1
dwavebinarycsp==0.1.2
fasteners==0.16.3
Fiona==1.8.20
Flask==2.0.1
geopandas==0.9.0
homebase==1.0.1
idna==2.10
importlib-metadata==4.5.0
importlib-resources==5.1.4
isort==5.8.0
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
matplotlib==3.4.2
mccabe==0.6.1
minorminer==0.2.5
munch==2.5.0
networkx==2.5.1
numpy==1.19.5
ortools==8.2.8710
pandas==1.2.4
penaltymodel==0.16.4
penaltymodel-cache==0.4.3
penaltymodel-lp==0.1.4
penaltymodel-mip==0.2.4
Pillow==8.2.0
plucky==0.4.3
protobuf==3.17.3
pylint==2.8.3
pyparsing==2.4.7
pyproj==3.1.0
pyqubo==1.0.10
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
Rtree==0.9.7
scipy==1.6.3
Shapely==1.7.1
six==1.16.0
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.4.1
